Penambahan izin route-route untuk admin dan user (progress 50%)

+++ routes/web.php
@@ -12,11 +12,11 @@ Route::post('/register', [AuthController::class, 'register']);
 
 Route::get('/dasbor', function () {
     return view('pages.dasbor');
-});
+})->middleware('role:Admin,User');
 
-Route::get('/resident', [ResidentController::class, 'index']);
-Route::get('/resident/create', [ResidentController::class, 'create']);
-Route::get('/resident/{id}', [ResidentController::class, 'edit']);
-Route::post('/resident', [ResidentController::class, 'store']);
-Route::put('/resident/{id}', [ResidentController::class, 'update']);
-Route::delete('/resident/{id}', [ResidentController::class, 'destroy']);
+Route::get('/resident', [ResidentController::class, 'index'])->middleware('role:Admin');
+Route::get('/resident/create', [ResidentController::class, 'create'])->middleware('role:Admin');
+Route::get('/resident/{id}', [ResidentController::class, 'edit'])->middleware('role:Admin');
+Route::post('/resident', [ResidentController::class, 'store'])->middleware('role:Admin');
+Route::put('/resident/{id}', [ResidentController::class, 'update'])->middleware('role:Admin');
+Route::delete('/resident/{id}', [ResidentController::class, 'destroy'])->middleware('role:Admin');
